refactor(ir): split materializeSession into its own file
Move materializeSessionToResolvedIR + the synthetic-top-level builder,
gateable-wrap helpers, and ExprNode→ResolvedExpr translator out of
compile_session.ts into compiler/ir/materialize_session.ts. compile_session.ts
keeps only the JIT entry point (compileSession + paramHandle stitching).

interpret_resolved.ts now imports materializeSessionToResolvedIR from the
new module and no longer pulls FlatPlan / compileResolved transitively
through the JIT-side bookend.
